+++ packages/nimble-components/src/label-provider/table/label-tokens.ts
@@ -21,7 +21,39 @@ export const tableCellActionMenuLabel = DesignToken.create<string>({
     cssCustomPropertyName: null
 }).withDefault(tableLabelDefaults.tableCellActionMenuLabel);
 
-export const tableColumnHeaderGroupedIndicatorLabel = DesignToken.create<string>({
-    name: 'table-column-header-grouped-indicator-label',
+export const tableColumnHeaderGroupedLabel = DesignToken.create<string>({
+    name: 'table-column-header-grouped-label',
     cssCustomPropertyName: null
-}).withDefault(tableLabelDefaults.tableColumnHeaderGroupedIndicatorLabel);
+}).withDefault(tableLabelDefaults.tableColumnHeaderGroupedLabel);
+
+export const tableColumnHeaderSortedAscendingLabel = DesignToken.create<string>(
+    {
+        name: 'table-column-header-sorted-ascending-label',
+        cssCustomPropertyName: null
+    }
+).withDefault(tableLabelDefaults.tableColumnHeaderSortedAscendingLabel);
+
+export const tableColumnHeaderSortedDescendingLabel = DesignToken.create<string>({
+    name: 'table-column-header-sorted-descending-label',
+    cssCustomPropertyName: null
+}).withDefault(tableLabelDefaults.tableColumnHeaderSortedDescendingLabel);
+
+export const tableSelectAllLabel = DesignToken.create<string>({
+    name: 'table-select-all-label',
+    cssCustomPropertyName: null
+}).withDefault(tableLabelDefaults.tableSelectAllLabel);
+
+export const tableGroupSelectAllLabel = DesignToken.create<string>({
+    name: 'table-group-select-all-label',
+    cssCustomPropertyName: null
+}).withDefault(tableLabelDefaults.tableGroupSelectAllLabel);
+
+export const tableRowSelectLabel = DesignToken.create<string>({
+    name: 'table-row-select-label',
+    cssCustomPropertyName: null
+}).withDefault(tableLabelDefaults.tableRowSelectLabel);
+
+export const tableRowOperationColumnLabel = DesignToken.create<string>({
+    name: 'table-row-operation-column-label',
+    cssCustomPropertyName: null
+}).withDefault(tableLabelDefaults.tableRowOperationColumnLabel);